requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.4.3
selenium==4.16.0
webdriver-manager==4.0.1